move list to it's own module
