+++ README.md
@@ -5,3 +5,4 @@ Experimental rust bindings for node-serialport using neon.
 TODO
 - [x] list (currently only `path` is supported)
 - [ ] all the other functions
+- [ ] make cross build work https://github.com/dcuddeback/libudev-sys#cross-compiling
